Move validators specs to their validators/ test directory
