+++ test/validators/member-of.spec.js
@@ -1,24 +1,23 @@
 import { expect } from 'chai';
-import validator from '../lib';
+import validator from '../../lib';
 
-describe('BeginWith validator', () => {
+describe('MemberOf validator', () => {
   const objectRules = {
     pet: [
-      {name: 'beginWith', params: [['dog', 'cat', 'fish']]}
+      {
+        name: 'memberOf', params: [['dog', 'cat', 'fish']]
+      }
     ]
   };
 
   it('should success for single item', () => {
     const acceptedInputs = [
       'dog',
-      'doge',
-      'dog kintamani',
-      'dog shiba inu',
-      'cat woman',
-      'fish erman'
+      'cat',
+      'fish'
     ];
 
-    acceptedInputs.forEach(acceptedInput => {
+    acceptedInputs.forEach(function test(acceptedInput) {
       const result = validator.validate(objectRules, {pet: acceptedInput});
       const err = result.messages;
 
@@ -30,11 +29,11 @@ describe('BeginWith validator', () => {
   it('should success for array item', () => {
     const acceptedInputs = [
       [],
-      ['dog', 'catherine'],
-      ['dog', 'doge', 'dog kintamani', 'dog shiba inu', 'cat woman', 'fish arowana']
+      ['fish'],
+      ['dog', 'cat']
     ];
 
-    acceptedInputs.forEach(acceptedInput => {
+    acceptedInputs.forEach(function test(acceptedInput) {
       const result = validator.validate(objectRules, {pet: acceptedInput});
       const err = result.messages;
 
@@ -43,60 +42,53 @@ describe('BeginWith validator', () => {
     });
   });
 
-  it('should fail for single item not beginning with string in list', () => {
+  it('should fail for single item not in the list', () => {
     const rejectedInputs = [
-      '',
-      'd',
-      'do',
-      'dug',
-      'dug up',
-      'cow',
-      'fisx',
+      'shark',
       null,
       undefined
     ];
 
-    rejectedInputs.forEach(rejectedInput => {
+    rejectedInputs.forEach(function test(rejectedInput) {
       const result = validator.validate(objectRules, {pet: rejectedInput});
       const err = result.messages;
 
       expect(result.success).to.equal(false);
       expect(err).to.have.property('pet');
-      expect(err.pet['beginWith:$1']).to.equal('Pet must begin with one of dog,cat,fish.');
+      expect(err.pet['memberOf:$1']).to.equal('Pet must be one of dog,cat,fish.');
     });
   });
 
-  it('should fail for array item which contains item not beginning with the string in list', () => {
+  it('should fail for array item which contains item not in the list', () => {
     const rejectedInputs = [
-      ['doge', 'dug'],
-      ['dog kintamani', 'd'],
-      ['dog kintamani', 'do']
+      ['dog', 'cat', 'fish', 'bear'],
+      ['shark']
     ];
 
-    rejectedInputs.forEach(rejectedInput => {
+    rejectedInputs.forEach(function test(rejectedInput) {
       const result = validator.validate(objectRules, {pet: rejectedInput});
       const err = result.messages;
 
       expect(result.success).to.equal(false);
       expect(err).to.have.property('pet');
-      expect(err.pet['beginWith:$1']).to.equal('Pet must begin with one of dog,cat,fish.');
+      expect(err.pet['memberOf:$1']).to.equal('Pet must be one of dog,cat,fish.');
     });
   });
 
   it('should fail for non string or non array of strings', () => {
     const rejectedInputs = [
-      ['doge', 'cat', {wild: 'object'}],
-      ['dog', 'catherine', 123],
+      ['dog', 'cat', {wild: 'object'}],
+      ['dog', 'cat', 123],
       {wild: 'object'}
     ];
 
-    rejectedInputs.forEach(rejectedInput => {
+    rejectedInputs.forEach(function test(rejectedInput) {
       const result = validator.validate(objectRules, {pet: rejectedInput});
       const err = result.messages;
 
       expect(result.success).to.equal(false);
       expect(err).to.have.property('pet');
-      expect(err.pet['beginWith:$1']).to.equal('Pet must begin with one of dog,cat,fish.');
+      expect(err.pet['memberOf:$1']).to.equal('Pet must be one of dog,cat,fish.');
     });
   });
 });
