+++ test/validators/min-length.spec.js
@@ -1,29 +1,22 @@
 import { expect } from 'chai';
-import validator from '../lib';
+import validator from '../../lib';
 
-describe('MaxLength validator', () => {
+describe('MinLength validator', () => {
   const ruleWithParam = {
-    name: ['maxLength:5']
+    name: ['minLength:5']
   };
 
   const acceptedInputs = [
-    'tele',
-    'phone',
-    '',
-    ' woi ',
-    '4l@y.',
-    12345,
-    123,
-    0,
-    -1234
+    'abcde',
+    'abcd ',
+    'wololo',
+    'pneumonoultramicroscopicsilicovolcanoconiosis'
   ];
 
   const rejectedInputs = [
-    '      ',
-    'asdasd',
-    'pneumonoultramicroscopicsilicovolcanoconiosis',
-    123456,
-    -12345,
+    '',
+    'asd',
+    'tele',
     null,
     undefined
   ];
@@ -45,7 +38,7 @@ describe('MaxLength validator', () => {
 
       expect(result.success).to.equal(false);
       expect(err).to.have.property('name');
-      expect(err.name['maxLength:$1']).to.equal('Name must contain at most 5 character(s).');
+      expect(err.name['minLength:$1']).to.equal('Name must contain at least 5 character(s).');
     });
   });
 });
