+++ test/validators/non-blank.spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import validator from '../lib';
+import validator from '../../lib';
 
 describe('NonBlank validator', () => {
   const rules = {
