+++ test/validators/string.spec.js
@@ -1,13 +1,13 @@
 import { expect } from 'chai';
-import validator from '../lib';
+import validator from '../../lib';
 
-describe('Not equal validator', () => {
+describe('String validator', () => {
   const rules = {
-    someField: ['not-equal:ya-shall-nat-pazz']
+    someField: ['string']
   };
 
   it('should success with valid input', () => {
-    const result = validator.validate(rules, {someField: 'pazz'});
+    const result = validator.validate(rules, {someField: 'yoyoyo'});
     const err = result.messages;
 
     expect(result.success).to.equal(true);
@@ -15,12 +15,11 @@ describe('Not equal validator', () => {
   });
 
   it('should fail with invalid input', () => {
-    const result = validator.validate(rules, {someField: 'ya-shall-nat-pazz'});
+    const result = validator.validate(rules, {someField: 123});
     const err = result.messages;
 
     expect(result.success).to.equal(false);
     expect(err).to.have.property('someField');
-    expect(err.someField['not-equal:$1'])
-      .to.equal('Some Field must not equal to ya-shall-nat-pazz.');
+    expect(err.someField.string).to.equal('Some Field is not a string.');
   });
 });
